requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2